+++ src/App.tsx
@@ -1,6 +1,6 @@
 import Playground from "./view/playground/index"
 import './App.css'
-import NotificationModal from "./components/NotificationModal"
+import NotificationModal from "./components/NotificationModal/NotificationModal"
 
 function App() {
